Extract HTML email conversion to plain text to an external service.
Move DefaultEmailSenderService.ConvertToPlainText to DefaultHtmlToPlainTextEmailBodyConverter; an implementation of IHtmlToPlainTextEmailBodyConverter.
