+++ src/ForEvolve.AspNetCore/Services/Emails/IHtmlToPlainTextEmailBodyConverter.cs
@@ -1,0 +1,7 @@
+﻿namespace ForEvolve.AspNetCore.Services
+{
+    public interface IHtmlToPlainTextEmailBodyConverter
+    {
+        string ConvertToPlainText(string body);
+    }
+}
